\name{formatSOAPParameter}
\alias{formatSOAPParameter}
\title{Formats an input string -- which is assumed to be NA, by default -- as a SOAP comopatible string}
\usage{
  formatSOAPParameter(par, makeList = TRUE)
}
\arguments{
  \item{par}{the parameter to format}

  \item{makeList}{whether or not to make the output into a
  list}
}
\value{
  the formatted parameter
}
\description{
  Formats an input string -- which is assumed to be NA, by
  default -- as a SOAP comopatible string
}
\author{
  Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}
}

