\name{pm.getProbesByGene}
\alias{pm.getProbesByGene}
\title{Get the probes associated with a specific gene...}
\usage{pm.getProbesByGene(conn, entrezID, platformID, authorityID)}
\description{Get the probes associated with a specific gene}
\value{a list containing information about the gene specified and a data.grame detailing the probes with which this 
gene is associated. There is a normalized [0-1] value representing the strength of that association according to each 
authority in the database.}
\author{Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}}
\arguments{\item{conn}{the connection object generated by pm.connect()}
\item{entrezID}{the EntrezID of the gene of interest.}
\item{platformID}{a number or vector of numbers representing the ID(s) of the platforms from which you want to find 
probes.}
\item{authorityID}{a number of vector of numbers representing the ID(s) of the authorities (sources) which you want 
to use to map this gene to probes. At the time of writing, 1 = BLAST, 2 = Vendor, 3 = Bioconductor.}}
